+++ src/state_Pattern/ContaState.java
@@ -1,8 +1,8 @@
 package state_Pattern;
 
 public interface ContaState {
-	void Saque(double valor);
-    void Deposito(double valor);
+	void Saque(int valor);
+    void Deposito(int valor);
     void Status();
     
 }
