+++ src/state_Pattern/SaldoPositivo.java
@@ -11,7 +11,7 @@ public class SaldoPositivo implements ContaState {
 		
 	}
 	
-	public void Deposito( double valor)
+	public void Deposito( int valor)
 	{
 		this.conta.saldo+=valor;
 		showMessageDialog(null,"Foi depositado R$ "+ valor);
@@ -28,7 +28,7 @@ public class SaldoPositivo implements ContaState {
 		}
 	}
 	
-	public void Saque(double valor)
+	public void Saque(int valor)
 	{
 		this.conta.saldo -= valor;
 		showMessageDialog(null,"Foi sacado R$ "+ valor);
